requirements.txt
requests==2.28.1
python-dotenv==1.0.0
urllib3==1.26.12
python-telegram-bot==20.3
aiohttp==3.8.1
colorama==0.4.6